requirements.txt
streamlit>=1.25.0
faiss-cpu>=1.7.4
numpy>=1.25.0
pandas>=2.1.0
Pillow>=10.0.0
torch>=2.1.0
transformers>=4.41.0
requests>=2.31.0